requirements.txt
Flask==3.0.0
werkzeug==3.0.1
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.7
python-dotenv==1.0.0
pystray==0.19.5
Pillow==10.1.0
pyinstaller==6.3.0
